+++ src/app/Home/page.js
@@ -2,7 +2,7 @@
 
 import { La_Belle_Aurore } from "next/font/google";
 import React, { useState } from "react";
-import NavBar from "../_components/_navbar.jsx/index.js";
+import NavBar from "../_components/NavBar.jsx/index.js";
 import Header from "../_components/_header.jsx";
 import BarChart from "../_components/BarChart.js";
 import LineChart from "../_components/LineChart";
